pyproject.toml
[tool.poetry]
name = "event-manager-test-assignment"
version = "0.1.0"
description = ""
authors = ["HCodeKeeper <ivannikovmax04@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
django-environ = "^0.11.2"
django-cors-headers = "^4.4.0"
psycopg2 = "^2.9.9"
faker = "^26.0.0"
celery = "^5.4.0"
redis = "^5.0.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
commitizen = "^3.27.0"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
coverage = "^7.5.4"
pytest-django = "^4.8.0"
flake8-pyproject = "^1.2.3"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'src.api.settings.local'
python_files = ['tests.py', 'test_*.py', '*_tests.py']
pythonpath = '. src'

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version = "0.2.0a19"
version_scheme = "pep440"
version_provider = "poetry"
major_version_zero = true
version_files = ["pyproject.toml:version"]

[tool.flake8]
exclude = [
    "*/migrations/*",
]

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
